[Refactor]: remove local member data source
