+++ .env.example
@@ -10,9 +10,7 @@ GOOGLE_API_KEY=
 PAGE_ACCESS_TOKEN=
 FB_VERIFY_TOKEN=
 
-# Member API
-MEMBER_API_URL=
-MEMBER_API_KEY=
+# Member MCP
 MEMBER_MCP_URL=https://ntuai.dev/api/mcp
 MEMBER_MCP_API_KEY=
 MEMBER_MCP_TIMEOUT_SECONDS=20
