+++ src/ian/config.py
@@ -59,10 +59,6 @@ FB_VERIFY_TOKEN = _env("FB_VERIFY_TOKEN")
 LINE_CHANNEL_ACCESS_TOKEN = _env("LINE_CHANNEL_ACCESS_TOKEN")
 LINE_CHANNEL_SECRET = _env("LINE_CHANNEL_SECRET")
 
-MEMBER_API_URL = _env("MEMBER_API_URL")
-MEMBER_API_KEY = _env("MEMBER_API_KEY")
-MEMBER_DB_FILE = DATA_DIR / "member_db.json"
-MEMBER_MAPPING_FILE = DATA_DIR / "member_mapping.csv"
 MEMBER_MCP_URL = _env("MEMBER_MCP_URL", "https://ntuai.dev/api/mcp")
 MEMBER_MCP_API_KEY = _env("MEMBER_MCP_API_KEY")
 MEMBER_MCP_TIMEOUT_SECONDS = _env_int("MEMBER_MCP_TIMEOUT_SECONDS", 20)
