+++ src/ian/domain/members.py
@@ -28,23 +28,7 @@ from pydantic import BaseModel, ConfigDict, Field, field_validator
 from ian.config import TZ_TPE
 
 
-TIER_ROLE_MAP = {
-    "STAFF": "幹部",
-    "VIP": "VIP 社員",
-}
-
-PLATFORM_FIELD_MAP = {
-    "Discord": "discord_acc_id",
-    "FB": "fb_acc_id",
-    "LINE": "line_acc_id",
-}
-
-VALID_SUBSCRIBE_PLATFORMS = {"discord", "fb", "line"}
-SUBSCRIBE_PLATFORM_FIELD = {
-    "discord": "discord_acc_id",
-    "fb": "fb_acc_id",
-    "line": "line_acc_id",
-}
+_VALID_SUBSCRIBE_PLATFORMS = {"discord", "fb", "line"}
 PERSONAL_PROMPT_MAX_LEN = 100
 SUBSCRIBE_PLATFORM_ORDER = ("discord", "fb", "line")
 
@@ -117,7 +101,7 @@ def normalize_subscribe(value: str | None) -> str | None:
         raise MemberDataError("subscribe 不可包含空的平台")
 
     invalid = sorted(
-        {part for part in normalized if part not in VALID_SUBSCRIBE_PLATFORMS}
+        {part for part in normalized if part not in _VALID_SUBSCRIBE_PLATFORMS}
     )
     if invalid:
         raise MemberDataError(f"不支援的訂閱平台: {', '.join(invalid)}")
@@ -195,40 +179,8 @@ class User(BaseModel):
         return tuple(Platform(part) for part in self.subscribe.split(","))
 
 
-def is_valid_member(member: dict, now: datetime | None = None) -> bool:
-    valid_date_str = member.get("valid_date", "")
-    if not valid_date_str:
-        return False
-    try:
-        valid_date = datetime.fromisoformat(valid_date_str.replace("Z", "+00:00"))
-        current = now or datetime.now(TZ_TPE)
-        return current <= valid_date
-    except (ValueError, TypeError):
-        return False
-
-
-def get_role_from_tier(tier: str) -> str:
-    if not tier:
-        return "社員"
-    return TIER_ROLE_MAP.get(tier, tier)
-
-
-def platform_field(platform: str) -> str | None:
-    return PLATFORM_FIELD_MAP.get(platform)
-
-
 def normalize_email(email: str) -> str:
     return email.strip().lower()
-
-
-def parse_subscribe_platforms(subscribe_str: str) -> list[str]:
-    raw_platforms = [p.strip().lower() for p in subscribe_str.split(",") if p.strip()]
-    return sorted(set(raw_platforms))
-
-
-def invalid_subscribe_platforms(subscribe_str: str) -> list[str]:
-    raw_platforms = [p.strip().lower() for p in subscribe_str.split(",") if p.strip()]
-    return [p for p in raw_platforms if p not in VALID_SUBSCRIBE_PLATFORMS]
 
 
 def normalize_personal_prompt(prompt_text: str) -> str:
